services/fan-pulse-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiokafka==0.10.0
redis==5.0.4
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
httpx==0.27.0
numpy==1.26.0
scikit-learn==1.4.2
pydantic==2.7.0
python-dotenv==1.0.1
bcrypt==4.1.3
faker==25.0.0
